+++ internal/configs/config.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/hashicorp/terraform/internal/addrs"
 	"github.com/hashicorp/terraform/internal/depsfile"
-	"github.com/hashicorp/terraform/internal/getproviders"
+	"github.com/hashicorp/terraform/internal/getproviders/providerreqs"
 )
 
 // A Config is a node in the tree of modules within a configuration.
@@ -92,7 +92,7 @@ type ModuleRequirements struct {
 	Name         string
 	SourceAddr   addrs.ModuleSource
 	SourceDir    string
-	Requirements getproviders.Requirements
+	Requirements providerreqs.Requirements
 	Children     map[string]*ModuleRequirements
 	Tests        map[string]*TestFileModuleRequirements
 }
@@ -100,7 +100,7 @@ type ModuleRequirements struct {
 // TestFileModuleRequirements maps the runs for a given test file to the module
 // requirements for that run block.
 type TestFileModuleRequirements struct {
-	Requirements getproviders.Requirements
+	Requirements providerreqs.Requirements
 	Runs         map[string]*ModuleRequirements
 }
 
@@ -297,14 +297,14 @@ func (c *Config) VerifyDependencySelections(depLocks *depsfile.Locks) []error {
 			lock = depLocks.Provider(providerAddr)
 		}
 		if lock == nil {
-			log.Printf("[TRACE] Config.VerifyDependencySelections: provider %s has no lock file entry to satisfy %q", providerAddr, getproviders.VersionConstraintsString(constraints))
+			log.Printf("[TRACE] Config.VerifyDependencySelections: provider %s has no lock file entry to satisfy %q", providerAddr, providerreqs.VersionConstraintsString(constraints))
 			errs = append(errs, fmt.Errorf("provider %s: required by this configuration but no version is selected", providerAddr))
 			continue
 		}
 
 		selectedVersion := lock.Version()
-		allowedVersions := getproviders.MeetingConstraints(constraints)
-		log.Printf("[TRACE] Config.VerifyDependencySelections: provider %s has %s to satisfy %q", providerAddr, selectedVersion.String(), getproviders.VersionConstraintsString(constraints))
+		allowedVersions := providerreqs.MeetingConstraints(constraints)
+		log.Printf("[TRACE] Config.VerifyDependencySelections: provider %s has %s to satisfy %q", providerAddr, selectedVersion.String(), providerreqs.VersionConstraintsString(constraints))
 		if !allowedVersions.Has(selectedVersion) {
 			// The most likely cause of this is that the author of a module
 			// has changed its constraints, but this could also happen in
@@ -313,8 +313,8 @@ func (c *Config) VerifyDependencySelections(depLocks *depsfile.Locks) []error {
 			// distinguish those cases here in order to avoid the more
 			// specific error message potentially being a red herring in
 			// the edge-cases.
-			currentConstraints := getproviders.VersionConstraintsString(constraints)
-			lockedConstraints := getproviders.VersionConstraintsString(lock.VersionConstraints())
+			currentConstraints := providerreqs.VersionConstraintsString(constraints)
+			lockedConstraints := providerreqs.VersionConstraintsString(lock.VersionConstraints())
 			switch {
 			case currentConstraints != lockedConstraints:
 				errs = append(errs, fmt.Errorf("provider %s: locked version selection %s doesn't match the updated version constraints %q", providerAddr, selectedVersion.String(), currentConstraints))
@@ -340,8 +340,8 @@ func (c *Config) VerifyDependencySelections(depLocks *depsfile.Locks) []error {
 //
 // If the returned diagnostics includes errors then the resulting Requirements
 // may be incomplete.
-func (c *Config) ProviderRequirements() (getproviders.Requirements, hcl.Diagnostics) {
-	reqs := make(getproviders.Requirements)
+func (c *Config) ProviderRequirements() (providerreqs.Requirements, hcl.Diagnostics) {
+	reqs := make(providerreqs.Requirements)
 	diags := c.addProviderRequirements(reqs, true, true)
 
 	return reqs, diags
@@ -352,8 +352,8 @@ func (c *Config) ProviderRequirements() (getproviders.Requirements, hcl.Diagnost
 //
 // If the returned diagnostics includes errors then the resulting Requirements
 // may be incomplete.
-func (c *Config) ProviderRequirementsShallow() (getproviders.Requirements, hcl.Diagnostics) {
-	reqs := make(getproviders.Requirements)
+func (c *Config) ProviderRequirementsShallow() (providerreqs.Requirements, hcl.Diagnostics) {
+	reqs := make(providerreqs.Requirements)
 	diags := c.addProviderRequirements(reqs, false, true)
 
 	return reqs, diags
@@ -366,7 +366,7 @@ func (c *Config) ProviderRequirementsShallow() (getproviders.Requirements, hcl.D
 // If the returned diagnostics includes errors then the resulting Requirements
 // may be incomplete.
 func (c *Config) ProviderRequirementsByModule() (*ModuleRequirements, hcl.Diagnostics) {
-	reqs := make(getproviders.Requirements)
+	reqs := make(providerreqs.Requirements)
 	diags := c.addProviderRequirements(reqs, false, false)
 
 	children := make(map[string]*ModuleRequirements)
@@ -380,7 +380,7 @@ func (c *Config) ProviderRequirementsByModule() (*ModuleRequirements, hcl.Diagno
 	tests := make(map[string]*TestFileModuleRequirements)
 	for name, test := range c.Module.Tests {
 		testReqs := &TestFileModuleRequirements{
-			Requirements: make(getproviders.Requirements),
+			Requirements: make(providerreqs.Requirements),
 			Runs:         make(map[string]*ModuleRequirements),
 		}
 
@@ -417,7 +417,7 @@ func (c *Config) ProviderRequirementsByModule() (*ModuleRequirements, hcl.Diagno
 // implementation, gradually mutating a shared requirements object to
 // eventually return. If the recurse argument is true, the requirements will
 // include all descendant modules; otherwise, only the specified module.
-func (c *Config) addProviderRequirements(reqs getproviders.Requirements, recurse, tests bool) hcl.Diagnostics {
+func (c *Config) addProviderRequirements(reqs providerreqs.Requirements, recurse, tests bool) hcl.Diagnostics {
 	var diags hcl.Diagnostics
 
 	// First we'll deal with the requirements directly in _our_ module...
@@ -435,7 +435,7 @@ func (c *Config) addProviderRequirements(reqs getproviders.Requirements, recurse
 			// don't exactly agree in practice 🙄 so this might produce new errors.
 			// TODO: Use the new parser throughout this package so we can get the
 			// better error messages it produces in more situations.
-			constraints, err := getproviders.ParseVersionConstraints(providerReqs.Requirement.Required.String())
+			constraints, err := providerreqs.ParseVersionConstraints(providerReqs.Requirement.Required.String())
 			if err != nil {
 				diags = diags.Append(&hcl.Diagnostic{
 					Severity: hcl.DiagError,
@@ -585,7 +585,7 @@ func (c *Config) addProviderRequirements(reqs getproviders.Requirements, recurse
 	return diags
 }
 
-func (c *Config) addProviderRequirementsFromProviderBlock(reqs getproviders.Requirements, provider *Provider) hcl.Diagnostics {
+func (c *Config) addProviderRequirementsFromProviderBlock(reqs providerreqs.Requirements, provider *Provider) hcl.Diagnostics {
 	var diags hcl.Diagnostics
 
 	fqn := c.Module.ProviderForLocalConfig(addrs.LocalProviderConfig{LocalName: provider.Name})
@@ -601,7 +601,7 @@ func (c *Config) addProviderRequirementsFromProviderBlock(reqs getproviders.Requ
 		// don't exactly agree in practice 🙄 so this might produce new errors.
 		// TODO: Use the new parser throughout this package so we can get the
 		// better error messages it produces in more situations.
-		constraints, err := getproviders.ParseVersionConstraints(provider.Version.Required.String())
+		constraints, err := providerreqs.ParseVersionConstraints(provider.Version.Required.String())
 		if err != nil {
 			diags = diags.Append(&hcl.Diagnostic{
 				Severity: hcl.DiagError,
