+++ internal/depsfile/locks.go
@@ -8,7 +8,7 @@ import (
 	"sort"
 
 	"github.com/hashicorp/terraform/internal/addrs"
-	"github.com/hashicorp/terraform/internal/getproviders"
+	"github.com/hashicorp/terraform/internal/getproviders/providerreqs"
 )
 
 // Locks is the top-level type representing the information retained in a
@@ -91,7 +91,7 @@ func (l *Locks) AllProviders() map[addrs.Provider]*ProviderLock {
 // non-lockable provider address then this function will panic. Use
 // function ProviderIsLockable to determine whether a particular provider
 // should participate in the version locking mechanism.
-func (l *Locks) SetProvider(addr addrs.Provider, version getproviders.Version, constraints getproviders.VersionConstraints, hashes []getproviders.Hash) *ProviderLock {
+func (l *Locks) SetProvider(addr addrs.Provider, version providerreqs.Version, constraints providerreqs.VersionConstraints, hashes []providerreqs.Hash) *ProviderLock {
 	if !ProviderIsLockable(addr) {
 		panic(fmt.Sprintf("Locks.SetProvider with non-lockable provider %s", addr))
 	}
@@ -175,7 +175,7 @@ func (l *Locks) SetSameOverriddenProviders(other *Locks) {
 // non-lockable provider address then this function will panic. Use
 // function ProviderIsLockable to determine whether a particular provider
 // should participate in the version locking mechanism.
-func NewProviderLock(addr addrs.Provider, version getproviders.Version, constraints getproviders.VersionConstraints, hashes []getproviders.Hash) *ProviderLock {
+func NewProviderLock(addr addrs.Provider, version providerreqs.Version, constraints providerreqs.VersionConstraints, hashes []providerreqs.Hash) *ProviderLock {
 	if !ProviderIsLockable(addr) {
 		panic(fmt.Sprintf("Locks.NewProviderLock with non-lockable provider %s", addr))
 	}
@@ -196,7 +196,7 @@ func NewProviderLock(addr addrs.Provider, version getproviders.Version, constrai
 	// assumes that we already sorted the items, which means that any duplicates
 	// will be consecutive in the sequence.
 	dedupeHashes := hashes[:0]
-	prevHash := getproviders.NilHash
+	prevHash := providerreqs.NilHash
 	for _, hash := range hashes {
 		if hash != prevHash {
 			dedupeHashes = append(dedupeHashes, hash)
@@ -315,9 +315,9 @@ func (l *Locks) Empty() bool {
 func (l *Locks) DeepCopy() *Locks {
 	ret := NewLocks()
 	for addr, lock := range l.providers {
-		var hashes []getproviders.Hash
+		var hashes []providerreqs.Hash
 		if len(lock.hashes) > 0 {
-			hashes = make([]getproviders.Hash, len(lock.hashes))
+			hashes = make([]providerreqs.Hash, len(lock.hashes))
 			copy(hashes, lock.hashes)
 		}
 		ret.SetProvider(addr, lock.version, lock.versionConstraints, hashes)
@@ -337,8 +337,8 @@ type ProviderLock struct {
 	// constraint but the previous selection still remains valid.
 	// "version" is therefore authoritative, while "versionConstraints" is
 	// just for a UI hint and not used to make any real decisions.
-	version            getproviders.Version
-	versionConstraints getproviders.VersionConstraints
+	version            providerreqs.Version
+	versionConstraints providerreqs.VersionConstraints
 
 	// hashes contains zero or more hashes of packages or package contents
 	// for the package associated with the selected version across all of
@@ -350,7 +350,7 @@ type ProviderLock struct {
 	// "h1:" for the first hash format version. Other hash versions following
 	// this scheme may come later. These versioned hash schemes are implemented
 	// in the getproviders package; for example, "h1:" is implemented in
-	// getproviders.HashV1 .
+	// providerreqs.HashV1 .
 	//
 	// There is also a legacy hash format which is just a lowercase-hex-encoded
 	// SHA256 hash of the official upstream .zip file for the selected version.
@@ -367,7 +367,7 @@ type ProviderLock struct {
 	// means we can only populate the hash for the current platform, and so
 	// it won't be possible to verify a subsequent installation of the same
 	// provider on a different platform.
-	hashes []getproviders.Hash
+	hashes []providerreqs.Hash
 }
 
 // Provider returns the address of the provider this lock applies to.
@@ -376,7 +376,7 @@ func (l *ProviderLock) Provider() addrs.Provider {
 }
 
 // Version returns the currently-selected version for the corresponding provider.
-func (l *ProviderLock) Version() getproviders.Version {
+func (l *ProviderLock) Version() providerreqs.Version {
 	return l.version
 }
 
@@ -388,7 +388,7 @@ func (l *ProviderLock) Version() getproviders.Version {
 // configuration have changed since a selection was made, and thus hint to the
 // user that they may need to run terraform init -upgrade to apply the new
 // constraints.
-func (l *ProviderLock) VersionConstraints() getproviders.VersionConstraints {
+func (l *ProviderLock) VersionConstraints() providerreqs.VersionConstraints {
 	return l.versionConstraints
 }
 
@@ -402,7 +402,7 @@ func (l *ProviderLock) VersionConstraints() getproviders.VersionConstraints {
 // of which must match in order for verification to be considered successful.
 //
 // Do not modify the backing array of the returned slice.
-func (l *ProviderLock) AllHashes() []getproviders.Hash {
+func (l *ProviderLock) AllHashes() []providerreqs.Hash {
 	return l.hashes
 }
 
@@ -437,6 +437,6 @@ func (l *ProviderLock) ContainsAll(target *ProviderLock) bool {
 //
 // At least one of the given hashes must match for a package to be considered
 // valud.
-func (l *ProviderLock) PreferredHashes() []getproviders.Hash {
-	return getproviders.PreferredHashes(l.hashes)
+func (l *ProviderLock) PreferredHashes() []providerreqs.Hash {
+	return providerreqs.PreferredHashes(l.hashes)
 }
