+++ internal/getproviders/providerreqs/hash_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: BUSL-1.1
 
-package getproviders
+package providerreqs
 
 import (
 	"testing"
